DO-955 Logo link + Copyrights
Added link to Landing on platform login page. https://razom.atlassian.net/browse/DO-955

Also according to DO-956, added Copyrights to platform login page. As i understand it should be changed, becuse page structure is incorrect(?). But if it acceptive as a temporary desicion, let's use it.
https://razom.atlassian.net/browse/DO-956

I was check everything few time, but anyway sure that there are something wrong, so if Copirights are unacceptable, i'll delete it.